int to long in rapl

+++ 09th-semester-EnergyComparer/EnergyComparer/Models/RaplDataPoint.cs
@@ -9,10 +9,10 @@ namespace EnergyComparer.Models
     public class RaplDataPoint
     {
         public DateTime Time { get; set; }
-        public int PSys { get; set; }
-        public int Core { get; set; }
-        public int Dram { get; set; }
-        public int Uncore { get; set; }
-        public int PackageZero { get; set; }
+        public long PSys { get; set; }
+        public long Core { get; set; }
+        public long Dram { get; set; }
+        public long Uncore { get; set; }
+        public long PackageZero { get; set; }
     }
 }
